chore: upgrade astro
