+++ package.json
@@ -10,13 +10,13 @@
     "astro": "astro"
   },
   "dependencies": {
-    "@astrojs/check": "0.9.4",
-    "@astrojs/mdx": "4.3.6",
-    "@astrojs/node": "9.4.4",
-    "@astrojs/rss": "4.0.12",
-    "@astrojs/solid-js": "5.1.1",
+    "@astrojs/check": "0.9.9",
+    "@astrojs/mdx": "5.0.6",
+    "@astrojs/node": "10.1.1",
+    "@astrojs/rss": "4.0.18",
+    "@astrojs/solid-js": "6.0.1",
     "@fontsource/fragment-mono": "^5.1.0",
-    "astro": "5.14.1",
+    "astro": "6.3.7",
     "cbor2": "^1.7.1",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
